+++ src/compas_timber/fabrication/__init__.py
@@ -18,12 +18,16 @@ from .mortise import Mortise
 from .slot import Slot
 from .pocket import Pocket
 from .free_contour import FreeContour
+from .text import Text
 from .btlx import TenonShapeType
 from .btlx import EdgePositionType
 from .btlx import LimitationTopType
+from .btlx import AlignmentType
 from .btlx import MachiningLimits
 from .btlx import StepShapeType
 from .btlx import BTLxFromGeometryDefinition
+from .btlx import Contour
+from .btlx import DualContour
 
 __all__ = [
     "BTLxWriter",
@@ -46,10 +50,14 @@ __all__ = [
     "Slot",
     "Pocket",
     "FreeContour",
+    "Text",
     "TenonShapeType",
     "EdgePositionType",
     "LimitationTopType",
+    "AlignmentType",
     "MachiningLimits",
     "StepShapeType",
     "BTLxFromGeometryDefinition",
+    "Contour",
+    "DualContour",
 ]