+++ tests/compas_timber/test_free_contour.py
@@ -1,4 +1,5 @@
 import pytest
+import math
 
 from compas.geometry import Point
 from compas.geometry import Polyline
@@ -7,6 +8,8 @@ from compas_timber.fabrication import BTLxWriter
 
 from compas_timber.elements import Plate
 from compas_timber.fabrication import FreeContour
+from compas_timber.fabrication import Contour
+from compas_timber.fabrication import DualContour
 from compas.data import json_loads
 from compas.data import json_dumps
 
@@ -14,12 +17,12 @@ from compas.data import json_dumps
 @pytest.fixture
 def plate():
     pline = Polyline([Point(0, 0, 0), Point(0, 200, 0), Point(100, 200, 0), Point(100, 0, 0), Point(0, 0, 0)])
-    return Plate(pline, 10.0)
+    return Plate.from_outline_thickness(pline, 10.0)
 
 
 def test_plate_blank():
     pline = Polyline([Point(0, 0, 0), Point(0, 200, 0), Point(100, 200, 0), Point(100, 0, 0), Point(0, 0, 0)])
-    plate = Plate(pline, 10.0)
+    plate = Plate.from_outline_thickness(pline, 10.0)
 
     assert len(plate.features) == 1
     assert isinstance(plate.features[0], FreeContour)
@@ -30,7 +33,7 @@ def test_plate_blank():
 
 def test_plate_blank_reversed():
     pline = Polyline([Point(0, 0, 0), Point(100, 0, 0), Point(100, 200, 0), Point(0, 200, 0), Point(0, 0, 0)])
-    plate = Plate(pline, 10.0)
+    plate = Plate.from_outline_thickness(pline, 10.0)
 
     assert len(plate.features) == 1
     assert isinstance(plate.features[0], FreeContour)
@@ -41,7 +44,7 @@ def test_plate_blank_reversed():
 
 def test_plate_blank_extension():
     pline = Polyline([Point(0, 0, 0), Point(0, 200, 0), Point(100, 200, 0), Point(100, 0, 0), Point(0, 0, 0)])
-    plate = Plate(pline, 10.0, blank_extension=5.0)
+    plate = Plate.from_outline_thickness(pline, 10.0, blank_extension=5.0)
 
     assert len(plate.features) == 1
     assert isinstance(plate.features[0], FreeContour)
@@ -53,10 +56,10 @@ def test_plate_blank_extension():
 def test_plate_contour():
     pline = Polyline([Point(0, 0, 0), Point(0, 200, 0), Point(100, 200, 0), Point(100, 0, 0), Point(0, 0, 0)])
     thickness = 10.0
-    plate = Plate(pline, thickness)
+    plate = Plate.from_outline_thickness(pline, thickness)
 
     expected = {
-        "header_attributes": {"ToolID": "0", "Name": "FreeContour", "ToolPosition": "right", "ReferencePlaneID": "4", "CounterSink": "no", "Process": "yes"},
+        "header_attributes": {"ToolID": "0", "Name": "FreeContour", "ToolPosition": "left", "ReferencePlaneID": "2", "CounterSink": "no", "Process": "yes"},
         "contour_attributes": {"Inclination": "0", "DepthBounded": "no", "Depth": "10.0"},
         "contour_points": [
             {"StartPoint": {"Y": "105.000", "X": "5.000", "Z": "0.000"}},
@@ -68,21 +71,21 @@ def test_plate_contour():
     }
 
     assert plate.features[0].params.header_attributes == expected["header_attributes"]
-    assert plate.features[0].params.as_dict()["Contour"].depth == thickness
+    assert TOL.is_close(plate.features[0].params.as_dict()["Contour"].depth, thickness)
 
 
 def test_plate_aperture_contour():
     plate_pline = Polyline([Point(0, 0, 0), Point(0, 200, 0), Point(100, 200, 0), Point(100, 0, 0), Point(0, 0, 0)])
     thickness = 10.0
     depth = 5.0
-    plate = Plate(plate_pline, thickness)
+    plate = Plate.from_outline_thickness(plate_pline, thickness)
     aperture_pline = Polyline([Point(25, 50, 0), Point(25, 150, 0), Point(75, 150, 0), Point(75, 50, 0), Point(25, 50, 0)])
-    contour = FreeContour.from_polyline_and_element(aperture_pline, plate, depth=depth)
+    contour = FreeContour.from_polyline_and_element(aperture_pline, plate, depth=depth, interior=True)
     plate.add_feature(contour)
 
     assert len(plate.features) == 2
     assert plate.features[1] == contour
-    assert contour.params.header_attributes["ToolPosition"] == "left"
+    assert contour.params.header_attributes["ToolPosition"] == "right"
     assert contour.params.header_attributes["CounterSink"] == "yes"
     assert contour.params.as_dict()["Contour"].depth == depth
 
@@ -91,16 +94,16 @@ def test_plate_aperture_contour_serialization():
     plate_pline = Polyline([Point(0, 0, 0), Point(0, 200, 0), Point(100, 200, 0), Point(100, 0, 0), Point(0, 0, 0)])
     thickness = 10.0
     depth = 5.0
-    plate = Plate(plate_pline, thickness)
+    plate = Plate.from_outline_thickness(plate_pline, thickness)
     aperture_pline = Polyline([Point(25, 50, 0), Point(25, 150, 0), Point(75, 150, 0), Point(75, 50, 0), Point(25, 50, 0)])
-    contour = FreeContour.from_polyline_and_element(aperture_pline, plate, depth=depth)
+    contour = FreeContour.from_polyline_and_element(aperture_pline, plate, depth=depth, interior=True)
 
     contour_copy = json_loads(json_dumps(contour))
     plate.add_feature(contour_copy)
 
     assert len(plate.features) == 2
     assert plate.features[1] == contour_copy
-    assert contour_copy.params.header_attributes["ToolPosition"] == "left"
+    assert contour_copy.params.header_attributes["ToolPosition"] == "right"
     assert contour_copy.params.header_attributes["CounterSink"] == "yes"
     assert contour_copy.params.as_dict()["Contour"].depth == depth
 
@@ -109,7 +112,7 @@ def test_plate_aperture_BTLx():
     plate_pline = Polyline([Point(0, 0, 0), Point(0, 200, 0), Point(100, 200, 0), Point(100, 0, 0), Point(0, 0, 0)])
     thickness = 10.0
     depth = 5.0
-    plate = Plate(plate_pline, thickness)
+    plate = Plate.from_outline_thickness(plate_pline, thickness)
     aperture_pline = Polyline([Point(25, 50, 0), Point(25, 150, 0), Point(75, 150, 0), Point(75, 50, 0), Point(25, 50, 0)])
     contour = FreeContour.from_polyline_and_element(aperture_pline, plate, depth=depth)
 
@@ -120,3 +123,77 @@ def test_plate_aperture_BTLx():
 
     assert processing_element.tag == "FreeContour"
     assert processing_element.attrib == contour.params.header_attributes
+
+
+def test_double_contour_plate():
+    pline_a = Polyline([Point(0, 0, 0), Point(0, 200, 0), Point(100, 200, 0), Point(100, 0, 0), Point(0, 0, 0)])
+    pline_b = Polyline([Point(-10, -10, 10), Point(-10, 210, 10), Point(110, 210, 10), Point(110, -10, 10), Point(-10, -10, 10)])
+    plate = Plate(pline_a, pline_b)
+
+    assert len(plate.features) == 1
+    assert isinstance(plate.features[0], FreeContour)
+    assert TOL.is_zero(plate.blank.xsize - 220.0)  # x-axis is the vector from `plate.outline[0]` to `plate.outline[1]`
+    assert TOL.is_zero(plate.blank.ysize - 120.0)
+    assert TOL.is_zero(plate.blank.zsize - 10.0)
+
+
+def test_contour_plate_blank():
+    pline_a = Polyline([Point(0, 0, 0), Point(0, 200, 0), Point(100, 200, 0), Point(100, 0, 0), Point(0, 0, 0)])
+    pline_b = Polyline([Point(-10, -10, 10), Point(-10, 210, 10), Point(110, 210, 10), Point(110, -10, 10), Point(-10, -10, 10)])
+    plate = Plate(pline_a, pline_b)
+
+    assert len(plate.features) == 1
+    assert isinstance(plate.features[0], FreeContour)
+    assert TOL.is_zero(plate.blank.xsize - 220.0)  # x-axis is the vector from `plate.outline[0]` to `plate.outline[1]`
+    assert TOL.is_zero(plate.blank.ysize - 120.0)
+    assert TOL.is_zero(plate.blank.zsize - 10.0)
+
+
+def test_contour_plate_simple_inclination():
+    pline_a = Polyline([Point(0, 0, 0), Point(0, 200, 0), Point(100, 200, 0), Point(100, 0, 0), Point(0, 0, 0)])
+    pline_b = Polyline([Point(-10, -10, 10), Point(-10, 210, 10), Point(110, 210, 10), Point(110, -10, 10), Point(-10, -10, 10)])
+    plate = Plate(pline_a, pline_b)
+
+    assert len(plate.features) == 1
+    assert isinstance(plate.features[0], FreeContour)
+    assert TOL.is_zero(plate.blank.xsize - 220.0)  # x-axis is the vector from `plate.outline[0]` to `plate.outline[1]`
+    assert TOL.is_zero(plate.blank.ysize - 120.0)
+    assert TOL.is_zero(plate.blank.zsize - 10.0)
+    assert TOL.is_close(plate.features[0].params.as_dict()["Contour"].inclination[0], -45.0)
+
+
+def test_contour_plate_multiple_inclination():
+    pline_a = Polyline([Point(0, 0, 0), Point(0, 200, 0), Point(100, 200, 0), Point(100, 0, 0), Point(0, 0, 0)])
+    pline_b = Polyline(
+        [
+            Point(-10, -10 * math.tan(math.pi / 6), 10),
+            Point(-10, 210, 10),
+            Point(110, 210, 10),
+            Point(110, -10 * math.tan(math.pi / 6), 10),
+            Point(-10, -10 * math.tan(math.pi / 6), 10),
+        ]
+    )
+    plate = Plate(pline_a, pline_b)
+
+    assert len(plate.features) == 1
+    assert isinstance(plate.features[0], FreeContour)
+    assert isinstance(plate.features[0].params.as_dict().get("Contour"), Contour)
+    assert TOL.is_allclose(plate.features[0].params.as_dict()["Contour"].inclination, [-45.0, -45.0, -45.0, -30.0])
+
+
+def test_dual_contour_plate():
+    pline_a = Polyline([Point(0, 0, 0), Point(0, 200, 0), Point(100, 200, 0), Point(100, 0, 0), Point(0, 0, 0)])
+    pline_b = Polyline(
+        [
+            Point(10, 10 * math.tan(math.pi / 6), 10),
+            Point(-10, 210, 10),
+            Point(110, 210, 10),
+            Point(110, -10 * math.tan(math.pi / 6), 10),
+            Point(10, 10 * math.tan(math.pi / 6), 10),
+        ]
+    )
+    plate = Plate(pline_a, pline_b)
+
+    assert len(plate.features) == 1
+    assert isinstance(plate.features[0], FreeContour)
+    assert isinstance(plate.features[0].params.as_dict().get("Contour"), DualContour)